system_module_2/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pip==21.3.1
setuptools==59.5.0
Werkzeug==2.0.2
wheel==0.37.0
google-cloud-datastore==2.4.0